refactor: Rename util to utils
